# RefreshGrid: async reading refs and logs in parallel