--- a/GitUI/CommandsDialogs/FormBrowse.cs
+++ b/GitUI/CommandsDialogs/FormBrowse.cs
@@ -568,9 +568,12 @@ namespace GitUI.CommandsDialogs
 
         private void UICommands_PostRepositoryChanged(object sender, GitUIEventArgs e)
         {
+            // Note that this called in most FormBrowse context to "be sure"
+            // that the repo has not been updated externally.
             ThreadHelper.JoinableTaskFactory.RunAsync(async () =>
             {
                 await this.SwitchToMainThreadAsync();
+
                 RefreshRevisions(e.GetRefs);
             }).FileAndForget();
 
@@ -1014,8 +1017,8 @@ namespace GitUI.CommandsDialogs
                 ThreadHelper.JoinableTaskFactory.RunAsync(async () =>
                 {
                     // Add a delay to not interfere with GUI updates when switching repository
-                    await Task.Delay(500);
                     await TaskScheduler.Default;
+                    await Task.Delay(500);
 
                     var result = Module.GetStashes(noLocks: true).Count;
 
@@ -2972,7 +2975,9 @@ namespace GitUI.CommandsDialogs
             var commitInfoPosition = AppSettings.CommitInfoPosition;
             if (commitInfoPosition == CommitInfoPosition.BelowList)
             {
+                CommitInfoTabControl.SelectedIndexChanged -= CommitInfoTabControl_SelectedIndexChanged;
                 CommitInfoTabControl.InsertIfNotExists(0, CommitInfoTabPage);
+                CommitInfoTabControl.SelectedIndexChanged += CommitInfoTabControl_SelectedIndexChanged;
                 CommitInfoTabControl.SelectedTab = CommitInfoTabPage;
 
                 RevisionsSplitContainer.FixedPanel = FixedPanel.Panel2;
@@ -2984,7 +2989,9 @@ namespace GitUI.CommandsDialogs
             {
                 // enough to fit CommitInfoHeader in most cases, when the width is (avatar + commit hash)
                 int width = DpiUtil.Scale(490) + SystemInformation.VerticalScrollBarWidth;
+                CommitInfoTabControl.SelectedIndexChanged -= CommitInfoTabControl_SelectedIndexChanged;
                 CommitInfoTabControl.RemoveIfExists(CommitInfoTabPage);
+                CommitInfoTabControl.SelectedIndexChanged += CommitInfoTabControl_SelectedIndexChanged;
 
                 if (commitInfoPosition == CommitInfoPosition.RightwardFromList)
                 {