Co-authored by RussKie
This replaces #9853
(changes pushed to #9853 too).

## Proposed changes

Improve the responsiveness in the loading phase and decrease
the revision loading time.

Revision info is read in two parallel steps:

1. Read current commit, refs, prepare grid etc.
2. Read all revisions with git-log.

Both these operations can take a long time and is done async.
Before part 1. was read synchrounously and blocked the interface
(git-log was async).
As the information is needed by step 2. when commits are added to the grid,
the information from step 1. is therefore protected with Lazy accessosors
and a semaphore.

RevisionReader (git-log) was previously also updating other information,
but is now a pure revision reader.

Other:

* The text label informing that loading of revisions still occur was
not displayed.

* NavigationHistory was normally always reset also if the grid was refreshed.
(With a potential race condition to keep existing history also if a new repo was opened).

* Clear side panel and the commit info panel at the beginning of the refresh.

* Avoid a Git call to get current head

* Optimizations for git-notes, assuming no notes exists.

## Screenshots 

Showing the difference for linux, limited number of refs so there are repos that benefits more
(after run in the debugger)

3.0 for loading refs is to load ahead/behind for the sidepanel, done in the background

The time differs from run to run, this is showing something I feel is representative.
The biggest difference will be seen in repos with many refs anyway.

The "sync" phase in the VS repo would be 10 s:
https://github.com/gitextensions/gitextensions/pull/9853#issuecomment-1030990082

### Before

170 ms in the sync phase (33 ms to load current HEAD)
Another 30 ms before logs are loaded
3.1 s to load revisions

3.3 s to load all

![image](https://user-images.githubusercontent.com/6248932/153774012-aee70db3-31fc-4b48-b2e9-4bee1e988bb2.png)

### After

![image](https://user-images.githubusercontent.com/6248932/153772934-b14c2fc3-56b5-436c-bc74-f05688b3263f.png)

(most of 196 ms after ls-files is time between activating and 
A few ms blocking the UI thread,
110 ms (a few more) in the spinner phase before loading refs, in this case the first rev is shown displayed just after that.
3.3s to load completely - git-log is more or less running from start to end

In this case the total loading time is actually a few ms more, but the git-log

![image](https://user-images.githubusercontent.com/6248932/153774396-c4e4f1c6-2f5c-48b8-9774-a3d280f28ad1.png)

![image](https://user-images.githubusercontent.com/6248932/153773051-b07c2280-3395-43f2-860f-2288cab62293.png)

## Test methodology <!-- How did you ensure quality? -->

Tests slightly updated, but mostly manual

## Merge strategy

I agree that the maintainer squash merge this PR (if the commit message is clear).

----

:black_nib: I contribute this code under [The Developer Certificate of Origin](../blob/master/contributors.txt).