@RussKie 2022-02-16
I'm observing an unexpected behaviour with this change - when I check change between commits (i.e., checkout a different one) in VS repo - the commit panel and the diff panels no longer update. Switching to another commit, selecting another commit or switching repos don't help. The only option is to restart the app. I am yet to find some time to debug this.