@gerhardol 2022-02-16
> I'm observing an unexpected behaviour with this change - when I check change between commits (i.e., checkout a different one) in VS repo - the commit panel and the diff panels no longer update. Switching to another commit, selecting another commit or switching repos don't help. The only option is to restart the app. I am yet to find some time to debug this.

Are you using the first commit, tabs was not updating there.

I will revert the reset of the side panel at refresh (but not when new repo is opened), that is just annoying.
The handling of _selectionTimer will be readdded, to Reset tabs (and commit info) while reading revisions.

I may add some other optimizations to indexwatcher etc (that is reset three times right now) - or in another PR if that fits better.
Some init is done many times. This is much less of a performance issue than any unnecessary Git command though.