@gerhardol 2022-02-18
A few more changes, more separated to #9871 (there will be a minor merge conflict.) These PRs belong together, maybe #9871 should be merged first.

I see one issue: The submodules are always collapsed at a refresh.