@gerhardol 2022-02-21
Squash and rebased to simplify review (#9833 and #9877 made some parallel changes).
See tmp/load-refs-parallel-to-log for previous commits.