@gerhardol 2022-02-26
The revgrid for submodules were not rendered until the super refs were displayed, which could take some seconds.

I have identified a few more optimizations for submodules.
If this can be merged first it would be great, but I want reviews to merge.