@RussKie 2022-02-26
Merge it 👍

On Sat, 26 Feb 2022, 12:03 pm Gerhard Olsson, ***@***.***>
wrote:

> The revgrid for submodules were not rendered until the super refs were
> displayed, which could take some seconds.
>
> I have identified a few more optimizations for submodules.
> If this can be merged first it would be great, but I want reviews to merge.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/gitextensions/gitextensions/pull/9864#issuecomment-1051409121>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABBTEXUUJ467DJ5RTLNCSVTU5ARGLANCNFSM5OJVZM2Q>
> .
> You are receiving this because you commented.Message ID:
> ***@***.***>
>