@gerhardol 2022-02-26
fixup! Lazy getUnfilteredRefs also when refreshed from grid

When refreshing from Browse getRefs is Lazy (shared with side panel).
This also uses a Lazy accessor when refreshing from grid
(refresh when changing could be slower than a refresh including side panel as several getRefs() was required)

Updated the comments

> Merge it 👍

Will check a detail for UpdateSelectedRef(), but will otherwise merge tomorrow.
(and add some related followups)