@mstv 2022-02-26
After merge conflicts and after (interactive) rebase, the RevisionGrid is empty since the recent changes. Before, only the graph was missing. Manual refresh works.