@gerhardol 2022-02-27
> After merge conflicts and after (interactive) rebase, the RevisionGrid is empty since the recent changes. Before, only the graph was missing. Manual refresh works.

I have seen it but cannot reproduce reliably, cannot reproduce better than #9223 (or #9792 that is closed but maybe describes the problem better).
Merging.