@RussKie 2022-02-28
🚀

On Mon, 28 Feb 2022, 9:46 am Gerhard Olsson, ***@***.***>
wrote:

> Merged #9864 <https://github.com/gitextensions/gitextensions/pull/9864>
> into master.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/gitextensions/gitextensions/pull/9864#event-6151063414>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABBTEXSPADIBWHHEHL7PLHLU5KSTVANCNFSM5OJVZM2Q>
> .
> You are receiving this because you commented.Message ID:
> ***@***.***>
>